ref: Move backtrace support out of sentry-core
