+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "sentry",
     "sentry-actix",
+    "sentry-backtrace",
     "sentry-core",
     "sentry-error-chain",
     "sentry-failure",
