+++ sentry-core/Cargo.toml
@@ -18,8 +18,7 @@ all-features = true
 
 [features]
 default = ["with_client_implementation", "with_device_info", "with_rust_info"]
-with_client_implementation = ["im", "url", "with_backtrace", "rand"]
-with_backtrace = ["backtrace", "regex"]
+with_client_implementation = ["im", "url", "rand"]
 with_debug_to_log = ["log"]
 with_device_info = ["libc", "hostname", "uname", "with_client_implementation"]
 with_rust_info = ["rustc_version", "with_client_implementation"]
@@ -27,13 +26,11 @@ with_debug_meta = ["findshlibs", "with_client_implementation"]
 with_test_support = []
 
 [dependencies]
-backtrace = { version = "0.3.44", optional = true }
 url = { version = "2.1.1", optional = true }
 log = { version = "0.4.8", optional = true, features = ["std"] }
 sentry-types = { path = "../sentry-types", version = "0.15.0" }
 env_logger = { version = "0.7.1", optional = true }
 lazy_static = "1.4.0"
-regex = { version = "1.3.4", optional = true }
 im = { version = "14.2.0", optional = true }
 libc = { version = "0.2.66", optional = true }
 hostname = { version = "0.3.0", optional = true }
