+++ sentry-core/src/lib.rs
@@ -80,7 +80,6 @@
 //! Default features:
 //!
 //! * `with_client_implementation`: Turns on the real client implementation.
-//! * `with_backtrace`: Enables backtrace support (automatically turned on in a few cases).
 //! * `with_device_info`: Enables the device info context.
 //! * `with_rust_info`: Enables the rust compiler info context instead of printing to stderr when
 //!   debug is enabled on the hub.
@@ -107,10 +106,6 @@ mod hub;
 mod scope;
 
 pub mod integrations;
-
-// NOTE: temporarily public, will move to `sentry-backtrace` soon
-#[cfg(feature = "with_backtrace")]
-pub mod backtrace_support;
 
 #[cfg(feature = "with_client_implementation")]
 mod client;
