+++ sentry-failure/Cargo.toml
@@ -13,7 +13,8 @@ Sentry Integration for failure crate
 edition = "2018"
 
 [dependencies]
-sentry-core = { version = "0.18.0", path = "../sentry-core", features = ["with_backtrace"] }
+sentry-core = { version = "0.18.0", path = "../sentry-core" }
+sentry-backtrace = { version = "0.18.0", path = "../sentry-backtrace" }
 lazy_static = "1.4.0"
 regex = "1.3.4"
 failure = "0.1.6"
