+++ sentry-log/Cargo.toml
@@ -13,7 +13,8 @@ Sentry Integration for log and env_logger crates
 edition = "2018"
 
 [dependencies]
-sentry-core = { version = "0.18.0", path = "../sentry-core", features = ["with_backtrace"] }
+sentry-core = { version = "0.18.0", path = "../sentry-core" }
+sentry-backtrace = { version = "0.18.0", path = "../sentry-backtrace" }
 log = { version = "0.4.8", features = ["std"] }
 env_logger = { version = "0.7.1", optional = true }
