+++ sentry-log/src/converters.rs
@@ -1,4 +1,4 @@
-use sentry_core::backtrace_support::current_stacktrace;
+use sentry_backtrace::current_stacktrace;
 use sentry_core::protocol::{Event, Exception};
 use sentry_core::{Breadcrumb, Level};
