+++ sentry/Cargo.toml
@@ -22,7 +22,7 @@ with_reqwest_transport = ["reqwest", "httpdate", "with_client_implementation"]
 with_curl_transport = ["curl", "httpdate", "serde_json", "with_client_implementation"]
 with_default_transport = ["with_reqwest_transport", "with_native_tls"]
 with_client_implementation = ["sentry-core/with_client_implementation"]
-with_backtrace = ["sentry-core/with_backtrace"]
+with_backtrace = ["sentry-backtrace"]
 with_panic = ["sentry-panic"]
 with_failure = ["sentry-failure"]
 with_debug_to_log = ["log", "sentry-core/with_debug_to_log"]
@@ -35,6 +35,7 @@ with_native_tls = ["reqwest/default-tls"]
 
 [dependencies]
 sentry-core = { version = "0.18.0", path = "../sentry-core", default-features = false }
+sentry-backtrace = { version = "0.18.0", path = "../sentry-backtrace", optional = true }
 sentry-failure = { version = "0.18.0", path = "../sentry-failure", optional = true }
 sentry-panic = { version = "0.18.0", path = "../sentry-panic", optional = true }
 log = { version = "0.4.8", optional = true, features = ["std"] }
